Fix: changed distance to be double type
- Updated euclideanDistance() method from CVRPClient to return a double instead of an int. And use Math.hypot internally
- Updated getCost() and getLastMoveCost() methods from Move to return a double
- Updated methods from IntrarouteSwap to conform the Move parent class
- Updated CVRPSolution totalDistance to be a double

+++ src/daa/project/crvp/moves/Move.java
@@ -26,13 +26,13 @@ public abstract class Move {
 	 * @return The difference in the objective function that applying the last move
 	 *         made
 	 */
-	public abstract int getLastMoveCost();
+    public abstract double getLastMoveCost();
 
 	/**
 	 * @return The objective function value of the solution made from applying the
 	 *         last move
 	 */
-	public abstract int getCost();
+    public abstract double getCost();
 
 	/**
 	 * @return Whether the solution made from applying the last move is feasible or
